+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - `Verik` annotation to enable inspections by IntelliJ plugin.
+- File annotation checker issues warnings if `Verik` annotation is missing.
 
 ## [0.1.10]
 ### Added